refactor: 사용자 지도 상호작용 이벤트 핸들링 및 상호작용 후 동작 로직 커스텀 훅으로 분리
